Revert "ecryptfs: forbid opening files without mmap handler"
commit 480546c9b2b8 ("ecryptfs: forbid opening

Files without mmap handler")

It fixed a local root exploit but also introduced a dependency on
the lower file system implementing an mmap operation just to open a file,
which is a bit of a heavy hammer.  The right fix is to have mmap depend
on the existence of the mmap handler instead.

Signed-off-by: Jeff Mahoney <jeffm@suse.com>
Cc: stable@vger.kernel.org
Signed-off-by: Tyler Hicks <tyhicks@canonical.com>
Git-repo:https://git.kernel.org/cgit/linux/kernel/git/stable/linux-stable.git
Git-commit: 78c4e172412de5d0456dc00d2b34050aa0b683b5
Signed-off-by: Srinivasa Rao Kuppala <srkupp@codeaurora.org>
Change-Id: I7c11ae0215d9b3a5944bb3be73d5423a4d008fcd